requirements.txt
PyContracts==1.6.3
decorator==3.4.0
gmpy2==2.0.3
ipdb==0.8
ipython==2.2.0
nose==1.3.3
pyparsing==2.0.2
